requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.28.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.44.0
django-cleanup==5.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-phonenumber-field==5.1.0
django-tastypie==0.14.3
django-widget-tweaks==1.4.8
djongo==1.3.6
dnspython==2.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
install==1.3.4
oauthlib==3.1.0
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2==2.8.6
py3-validate-email==1.0.3
pycparser==2.20
PyJWT==1.7.1
pymongo==3.12.0
python-dateutil==2.8.1
python-magic-bin==0.4.14
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.2.4
twilio==6.59.0
urllib3==1.26.5
whitenoise==5.2.0
